move delete family modal and action button to app directory
