+++ libs/design/Modal/index.ts
@@ -1,3 +1,2 @@
-export * from './DeleteFamilyModal';
 export * from './LogoutModal';
 export * from './ShowModal';
