+++ libs/design/index.ts
@@ -1,4 +1,3 @@
-export * from './ActionButton';
 export * from './Alert';
 export * from './Badge';
 export * from './Breadcrumbs';
